moved ClassLoaderUtils to core project since used by multiple projects
